Move bedspace utilities out of service
