+++ server/utils/bedspaceUtils.ts
@@ -1,9 +1,16 @@
-import { Cas3Bedspace, Cas3Premises } from '@approved-premises/api'
-import { BedspaceStatus, PageHeadingBarItem, PlaceContext } from '@approved-premises/ui'
+import {
+  Cas3Bedspace,
+  Cas3BedspaceArchiveAction,
+  Cas3BedspaceStatus,
+  Cas3Premises,
+  Characteristic,
+} from '@approved-premises/api'
+import { BedspaceStatus, PageHeadingBarItem, PlaceContext, SummaryList } from '@approved-premises/ui'
 import paths from '../paths/temporary-accommodation/manage'
 import { addPlaceContext } from './placeUtils'
-import { dateIsInFuture } from './dateUtils'
+import { DateFormats, dateIsInFuture } from './dateUtils'
 import config from '../config'
+import { convertToTitleCase } from './utils'
 
 export function bedspaceActions(
   premises: Cas3Premises,
@@ -133,3 +140,127 @@ export function setDefaultStartDate(userInput: Record<string, unknown>) {
     userInput['startDate-year'] = String(today.getFullYear())
   }
 }
+
+export const summaryList = (bedspace: Cas3Bedspace): SummaryList => {
+  const rows = [
+    {
+      key: { text: 'Bedspace status' },
+      value: { html: formatBedspaceStatus(bedspace) },
+    },
+    {
+      key: { text: 'Start date' },
+      value: { text: formatBedspaceDate(bedspace.startDate) },
+    },
+  ]
+
+  if (bedspace.archiveHistory && bedspace.archiveHistory.length > 0) {
+    rows.push({
+      key: { text: 'Archive history' },
+      value: {
+        html: formatArchiveHistory(bedspace.archiveHistory),
+      },
+    })
+  }
+
+  rows.push(
+    {
+      key: { text: 'Bedspace details' },
+      value: { html: formatBedspaceDetails(bedspace.characteristics) || 'None' },
+    },
+    {
+      key: { text: 'Additional bedspace details' },
+      value: { html: formatNotes(bedspace.notes ?? 'None') },
+    },
+  )
+  return {
+    rows,
+  }
+}
+
+const getBedspaceStatusTagColour = (status: Cas3BedspaceStatus): string => {
+  switch (status) {
+    case 'online':
+      return 'govuk-tag--green'
+    case 'archived':
+      return 'govuk-tag--grey'
+    case 'upcoming':
+      return 'govuk-tag--blue'
+    default:
+      return 'govuk-tag--grey'
+  }
+}
+
+const formatBedspaceStatus = (bedspace: Cas3Bedspace): string => {
+  const tagClass = getBedspaceStatusTagColour(bedspace.status)
+  let html = `<strong class="govuk-tag ${tagClass}">${convertToTitleCase(bedspace.status)}</strong>`
+
+  if (bedspace.status === 'online' && bedspace.endDate) {
+    html += `<br><span class="govuk-!-display-inline-block govuk-!-margin-top-2">Scheduled archive date ${formatBedspaceDate(bedspace.endDate)}</span>`
+  } else if (bedspace.status === 'archived' && new Date(bedspace.scheduleUnarchiveDate) > new Date()) {
+    html += `<br><span class="govuk-!-display-inline-block govuk-!-margin-top-2">Scheduled online date ${formatBedspaceDate(bedspace.scheduleUnarchiveDate)}</span>`
+  }
+
+  return html
+}
+
+const formatBedspaceDate = (dateString: string | undefined | null): string => {
+  if (dateString === undefined || dateString === null || dateString === '') {
+    return ''
+  }
+
+  return DateFormats.isoDateToUIDate(dateString)
+}
+
+const formatBedspaceDetails = (characteristics: Array<Characteristic>): string => {
+  return characteristics
+    .map(characteristic => `<span class="hmpps-tag-filters">${characteristic.name}</span>`)
+    .join(' ')
+}
+
+const formatArchiveHistory = (archiveHistory: Array<Cas3BedspaceArchiveAction>): string => {
+  return archiveHistory
+    .map(action => {
+      const verb = action.status === 'archived' ? 'Archive' : convertToTitleCase(action.status)
+      return `<div>${verb} date ${formatBedspaceDate(action.date)}</div>`
+    })
+    .join('')
+}
+
+export const summaryListForBedspaceStatus = (bedspace: Cas3Bedspace): SummaryList => {
+  let endDate = 'No end date added'
+
+  if (bedspace.endDate) {
+    endDate = DateFormats.isoDateToUIDate(bedspace.endDate)
+
+    if (bedspaceStatus(bedspace) === 'online') {
+      endDate += ` (${DateFormats.isoDateToDaysFromNow(bedspace.endDate)})`
+    }
+  }
+
+  return {
+    rows: [
+      {
+        key: textValue('Bedspace status'),
+        value: htmlValue(
+          bedspaceStatus(bedspace) === 'online'
+            ? `<span class="govuk-tag govuk-tag--green">Online</span>`
+            : `<span class="govuk-tag govuk-tag--grey">Archived</span>`,
+        ),
+      },
+      {
+        key: textValue('Bedspace end date'),
+        value: textValue(endDate),
+      },
+    ],
+  }
+}
+
+const textValue = (value: string) => {
+  return { text: value }
+}
+
+const htmlValue = (value: string) => {
+  return { html: value }
+}
+
+const formatNotes = (notes: string): string => notes.replace(/\n/g, '<br />')
